Re-enable the js bindings targets
These were temporarily disabled in the repo split because I was confused
about how they would be developed. We're going to develop these in the
mojo repo, so this re-enables the build targets and adds the tests to
the mojob set so they run.

R=hansmuller@chromium.org

Review URL: https://codereview.chromium.org/677773003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f2b3f41db09c6c54eef25347987f14e8b7472a9d