Pass command line args to apptests
Apptests should be able to access passed on the command line when they're run.
The test runner will now export the args and cause them to be handed to
the ApplicationImpl created as a part of the ApplicationTestBase class.

BUG=None
TEST=run app tests with apptest_runner.py
R=msw@chromium.org

Review URL: https://codereview.chromium.org/744973002

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: c15c806c42e5e97b90bb4816c7601c8ab58fa8ad