Move remaining test targets out of //mojo/public.
The test targets currently in //mojo/public depend on the EDK to run and thus
are not part of the public SDK. This CL converts the remaining test targets in
the public SDK into testonly mojo_sdk_source_sets, and adds new test targets
under the EDK that depend on the corresponding source_sets as well as the EDK
mechanism for running the unittests.

Note that the dependencies from the test targets to the source_sets are
specified via relative paths as part of the work toward having the EDK be
flexible wrt its location in a client repo.

Initial review: https://codereview.chromium.org/767083006/

BUG=439082
TBR=jamesr@chromium.org

Review URL: https://codereview.chromium.org/796853003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 0bcf40dec38480296394ffb28e62dac9a6854e3d