Allow ApplicationTestBase clients to avoid a default MessageLoop.
Provides a means for overriding the creation of a default message
loop.  Such a loop is unnecessary for many tests and causes issues
with the bevy of existing chrome tests which expect to be able to
create their own message loop.

BUG=none
TEST=https://codereview.chromium.org/806443002/
R=msw@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/792463007

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: dcc2f22a40eb6070ba79b322f8db3a2a7cc1bb22