Changes magic sequence to '!#mojo ' and strips 'mojo ' when building url
Prior to this change the magic sequence was '#!mojo:' and the url of the content handler app included mojo:. This works only for mojo: urls, but we want
other schemes. So, the magic sequence is now '#!mojo ' and 'mojo ' is not part of the resulting url.

R=aa@chromium.org

Review URL: https://codereview.chromium.org/821403003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 05833db1b922104838d2d7e128ff601b2b3461ec