Always construct ApplicationImpl using a ShellPtr, simplify apptest
The only sort of handle that an ApplicationImpl can successfully
be constructed with is a ShellPtr, so always construct it with that
static type. Also simplify the apptest startup code that is just trying
to eat the command line arguments to not spin up the production (!)
implementation of Application.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/868223002

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 553ee08efa5a33d6b02cb10bbec8f47633ced839