Making comments on MojoReadMessage a little easier to read.
The existing comment was pretty good but I did have to re-read some parts of it a couple times to get the right understanding. I had some confusion over some parts that looked like a division but were actually a slash. I was also a bit confused about which message was being referred to as the "next message" in some parts of the docs.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/863473004

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: d043808fd0bd58390c849d54be6dcac99eb9aae9