Allow adapting a lambda to a mojo::Callback
This allows a mojo::Callback to be implicitly converted from a lambda
or any other type with an appropriate operator() overload. This still
allows constructing from a type with an appriopriate Run() method, such
as a base::Callback<>. If a type provides both, this prefers binding to
the operator() overload. This implementation does not use any C++11
library features although the type detection would be nicer if it did.

Special care must be taken when binding to a lambda to think about the
lifecycle of all captured variables. In general, it's safe to bind a
callback from a mojom-generated interface to variables that have
lifetimes longer than the lifetime of the pipe driving the interface,
so common use cases are generally safe i.e.

class Foo {
  void Something() {
    ptr_->CallWithCallback(..., [&calls_] () { calls_++; });
  }

  MyInterfacePtr ptr_;
  int calls_;
};

is safe even though it binds a class member since the callback cannot
run after ptr_ is destroyed.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/884383003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 6d11a82aa083f8da67a7f8e7cb808a55ffd304d1