More tweaks to ServiceProvider
Having the factory return a core.Listener is a bit limiting (for
example, it prevents the factory from forwarding the request on, or
delaying use, or other interesting use cases). So, changes things so
it's assumed the factory takes ownership of the pipe and does whatever
it wants with it.

R=zra@google.com, zra@chromium.org

Review URL: https://codereview.chromium.org/914893003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 70bfb71dae6e4c38251e4e6e4bbbc5d3ff8f64d3