Add initializeFromShellProxy to application.dart
This CL adds the ability to initialize the Dart Application class with a shell
proxy (rather than with a MojoHandle). This change lets Sky use the Dart
bindings for Mojo.

This CL was reviewed on the SkyDart branch in
https://codereview.chromium.org/919883002/

R=zra@google.com, zra@chromium.org

Review URL: https://codereview.chromium.org/922083002

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 94a22312a24fb2ca262f3f9fb5b3ba377b29844b