Dart: Enables strict mode only when specified.
I'm not exactly sure what the right way to do this is. In this CL, if
the URL for a Dart script ends with ?strict, the Isolate that runs
that script (and Isolates that it spawns) is run in strict mode (that
is, with type-checking and assertions enabled.)

This change relies on [1] upstream, which I'm working on concurrently
with this change.

If this is not a good way to get parameters like "checked" to the
content handler, I'd be interested in learning about alternatives.

If this is okay, I'd be interested to learn whether there is already
code somewhere I can use to extract parameters from the URL.

[1] https://codereview.chromium.org/883263004

BUG=
R=aa@chromium.org, abarth@chromium.org

Review URL: https://codereview.chromium.org/867823010

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 092afbe9440d92ecd7af1491a9c8220290bd7dd0