Python promises: Allow to retrieve members of a promise
This CL updates promise to that one can get attributes of a promise, and
call a promises. Those will always return a promise that will resolved
to the result of the same operation on the result of the first promise.

For example, if we have a class Bar() with a method bar, and a promise p
that will resolve to an instance of Bar, this allows to call p.bar()
that will resolve to the result of the bar method on the resolved Bar
object.

R=blundell@chromium.org
BUG=

Review URL: https://codereview.chromium.org/960793006

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 8fedc59d5589f273ede473a3c27db826a33732e0