Dart: Better handle leak checks. close() is async.
This change adds tracking of MojoHandle creation and closing in Debug
mode, and makes it possible to emit a log of leaked handles on Mojo app
shutdown. If any handles are leaked, causing the finalizer to need to
close a handle, an error message is emitted with LOG(ERROR). All of the
handles created by an isolate are also closed on an unhandled exception.
This should prevent hangs.

To aid tracking handles, close() now returns a Future in most cases.
This is because a call to close() will generally send a command to the
handle watcher, and so the actual close() of the underlying handle
happens asynchronously. In particular when the future for a call to
close() on an Application resolves, we should be able to do:

assert(MojoHandle.reportLeakedHandles());

This change also fixes analyzer warnings and removes the special case
to ignore warnings from the analyzer's wrapper script.

BUG=
R=erg@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/996923003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: b7959afdc4fa4acb65c67cafa6985394633138c8