go/bindings: 8-byte align uint64 for atomic operations
It turned out that on ARM we should manually 8-byte align
uint64 values that are accessed atomically. See
http://golang.org/pkg/sync/atomic/#pkg-note-BUG
We do it by making a struct that has an uint64
non-exported and create an interface for it so that
it can't be a part of a struct and aligned wrong way.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1005683002

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 4a1dabbf487eafb9a07734e93ca53b6b661f0f22